requirements.txt
requests
lxml
feedparser
pyyaml